tools/python_jwt/requirements.txt
cryptography==43.0.1
PyJWT==2.9.0
